refactor : WinningNumber 검증 생성하기 전에 하도록 수정
